pyproject.toml
[project]
name = "aioscore"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6",
  "jsonschema>=4.22",
  "typer>=0.12",
  "PyYAML>=6.0",            # <— used by final.py
]

[project.scripts]
aioscore = "aioscore.cli.main:app"

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]
